requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
kafka-python-ng
groq
datadog-api-client
requests
pydantic
python-dotenv
aiofiles
python-multipart